requirements.txt
flask==3.0.0
werkzeug==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
pymysql==1.1.0